Remove js extension from imports; rename more files
