+++ moped-editor/src/utils/activityLogHelpers.jsx
