+++ moped-editor/src/utils/signalComponentHelpers.jsx
